requirements.txt
streamlit
langchain
openai
replicate
requests
transformers==4.40.2
accelerate==0.30.1
bitsandbytes==0.43.1
langchain==0.1.2
google.generativeai